Moved function to collection.py; added path parameter with overriding option; implemented one log function; Changed that its only getting EYS
